backend/requirements.txt
yfinance>=0.2.0
requests>=2.28.0
beautifulsoup4>=4.11.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
tradingeconomics==4.5.6
